+++ src/components/Login/Login.jsx
@@ -3,7 +3,7 @@ import {useEffect, useState} from 'react';
 import '../general/content.css';
 import './Login.css';
 import FormSection from '../FormSection/FormSection';
-import {useForm, useFormWithValidation} from '../ValidForm/ValidForm';
+import {useForm, useFormWithValidation} from '../../hooks/useValidForm.jsx';
 
 function Login({ handleLogin, errorApi }) { // @props из App.js - аутентификация пользователя
     /** логика и стейты: values.name, values.emaiil, ... --> in ValidForm component */
